Crate smol_concurrency_tools

smol-concurrency-tools

A collection of nice tools developed for projects using smol.

Documentation can be found here: loafey.se/smol-concurrency-tools.

Core content:

Interval: A Stream which runs infinitely and returns on a set interval.

Repeat: A stream which contains a task which will be repeated infinitely (i.e as long as it gets awaited).

Repeats: A stream which repeats a set of tasks infinitely (i.e as long as it gets awaited).

Macros:

repeat: Nice wrapper around Repeats

let mut streams = repeat!(
    async {
        Timer::after(Duration::from_secs(1)).await;
        1
    },
    async {
        Timer::after(Duration::from_secs(2)).await;
        2
    }
);
while let Some(p) = streams.next().await {
    println!("{p}");
}

select: Port of the popular select! macro from other async toolkits, enabled with the select feature:

loop {
    select!(
        (sec1.next(), |d| {
            let d = d.unwrap();
            println!(
                "1 second - since_last: {:0.5}s, total_time: {:0.5}s",
                d.as_secs_f32(),
                timer.elapsed().as_secs_f32()
            );
        }),
        (sec2.next(), |d| {
            let d = d.unwrap();
            println!(
                "2 second - since_last: {:0.5}s, total_time: {:0.5}s",
                d.as_secs_f32(),
                timer.elapsed().as_secs_f32()
            );
        })
    );
}

Modules

paste

Re-export of the paste crate.

results

Macros

repeat

Creates tasks that get re-created infinitely as long as they are being polled. Just a wrapper around Repeats and RepeatsBuilder.

select

A clone of the popular select! macro, but made compatible with smol. Requires the feature select and the nightly feature macro_metavar_expr. Disabled by default due to these requirements.

Structs

Interval

A Stream which runs infinitely and returns on a set interval.

Repeat

A Stream which contains a task which will be repeated infinitely (i.e as long as it gets awaited).

Repeats

Stream which repeats a set of tasks infinitely (i.e as long as it gets awaited).

RepeatsBuilder

A builder struct for Repeats.

Functions

repeats

Create a RepeatsBuilder.